Restructure project folder to have apps within project
Simplifying the project structure to have apps within the Django
project folder. Renamed the project to Panalytics and updated any
associated references. Renamed the analytics app to 'core' as it is
the core of the Panalytics project.
