+++ panalytics/core/admin.py
