+++ panalytics/core/models.py
