+++ panalytics/core/tests/test_model_pageview.py
@@ -33,8 +33,8 @@ def get_request_url_title_dimensions():
     return RequestFactory().get('/a.gif', query)
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_url_to_obj_create(
     mock_pageview_objects_create, mock_project_objects_get, get_request_url
 ):
@@ -47,8 +47,8 @@ def test_create_from_request_passes_url_to_obj_create(
     assert kwargs['url'] == 'http://example.com'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_default_referer_to_obj_create(
     mock_pageview_objects_create, mock_project_objects_get, get_request_url
 ):
@@ -58,8 +58,8 @@ def test_create_from_request_passes_default_referer_to_obj_create(
     assert kwargs['referer'] == ''
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_referer_to_obj_create(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -71,8 +71,8 @@ def test_create_from_request_passes_referer_to_obj_create(
     assert kwargs['referer'] == 'http://refer.com'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_source_as_referer_to_obj_create(
     mock_pageview_objects_create, mock_project_objects_get
 ):
@@ -84,8 +84,8 @@ def test_create_from_request_passes_source_as_referer_to_obj_create(
     assert kwargs['referer'] == 'email'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_referer_instead_or_source_to_obj_create(
     mock_pageview_objects_create, mock_project_objects_get
 ):
@@ -97,8 +97,8 @@ def test_create_from_request_passes_referer_instead_or_source_to_obj_create(
     assert kwargs['referer'] == 'http://refer.com'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_url_with_path_to_obj_create(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -113,8 +113,8 @@ def test_create_from_request_passes_url_with_path_to_obj_create(
     assert kwargs['url'] == 'http://example.com/about'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_correct_title_to_obj_create(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -126,8 +126,8 @@ def test_create_from_request_passes_correct_title_to_obj_create(
     assert kwargs['title'] == 'Test Title'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_default_title_to_obj_create(
     mock_pageview_objects_create, mock_project_objects_get, get_request_url
 ):
@@ -137,8 +137,8 @@ def test_create_from_request_passes_default_title_to_obj_create(
     assert kwargs['title'] == ''
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_correct_window_dimensions_to_obj_create(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -151,8 +151,8 @@ def test_create_from_request_passes_correct_window_dimensions_to_obj_create(
     assert kwargs['window_height'] == '675'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_default_window_dimensions_when_empty(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -165,8 +165,8 @@ def test_create_from_request_passes_default_window_dimensions_when_empty(
     assert kwargs['window_height'] == '0'
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_passes_unique_visit_to_obj_create(
     mock_pageview_objects_create,
     mock_project_objects_get,
@@ -178,8 +178,8 @@ def test_create_from_request_passes_unique_visit_to_obj_create(
     assert kwargs['unique_visit']
 
 
-@patch('analytics.models.Project.objects.get')
-@patch('analytics.models.PageView.objects.create')
+@patch('panalytics.core.models.Project.objects.get')
+@patch('panalytics.core.models.PageView.objects.create')
 def test_create_from_request_not_unique_visit_if_url_referer_domain_the_same(
     mock_pageview_objects_create, mock_project_objects_get
 ):
