+++ panalytics/core/tests/test_view_script.py
@@ -24,7 +24,7 @@ def script_get_request_no_tid():
     return RequestFactory().get('/a.js', content_type='text/javascript')
 
 
-@patch('analytics.models.Project.is_valid_tracking_id')
+@patch('panalytics.core.models.Project.is_valid_tracking_id')
 def test_script_view_200_status(mock_is_valid_tracking_id, script_get_request):
     mock_is_valid_tracking_id.return_value = True
     response = script(script_get_request)
@@ -32,7 +32,7 @@ def test_script_view_200_status(mock_is_valid_tracking_id, script_get_request):
     assert response.status_code == 200
 
 
-@patch('analytics.models.Project.is_valid_tracking_id')
+@patch('panalytics.core.models.Project.is_valid_tracking_id')
 def test_script_view_content_type(
     mock_is_valid_tracking_id, script_get_request
 ):
@@ -42,7 +42,7 @@ def test_script_view_content_type(
     assert response['content-type'] == 'text/javascript'
 
 
-@patch('analytics.models.Project.is_valid_tracking_id')
+@patch('panalytics.core.models.Project.is_valid_tracking_id')
 def test_script_returns_javascript_with_domain(
     mock_is_valid_tracking_id, script_get_request
 ):
@@ -52,7 +52,7 @@ def test_script_returns_javascript_with_domain(
     assert EXPECTED_JAVASCRIPT in str(response.content)
 
 
-@patch('analytics.models.Project.is_valid_tracking_id')
+@patch('panalytics.core.models.Project.is_valid_tracking_id')
 def test_script_view_raises_404_status_with_invalid_tid(
     mock_is_valid_tracking_id, script_get_request_no_tid
 ):
