+++ panalytics/core/urls.py
