+++ panalytics/core/views.py
