+++ panalytics/settings.py
@@ -1,15 +1,3 @@
-"""
-Django settings for analytics_project project.
-
-Generated by 'django-admin startproject' using Django 3.0.3.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.0/ref/settings/
-"""
-
 import os
 from decouple import config
 
@@ -33,7 +21,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'analytics'
+
+    'panalytics.core'
 ]
 
 MIDDLEWARE = [
@@ -46,7 +35,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'analytics_project.urls'
+ROOT_URLCONF = 'panalytics.urls'
 
 TEMPLATES = [
     {
@@ -64,7 +53,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'analytics_project.wsgi.application'
+WSGI_APPLICATION = 'panalytics.wsgi.application'
 
 
 # Database
