+++ panalytics/urls.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
 from django.urls import include, path
 
-from analytics import urls as analytics_urls
+from .core import urls as core_urls
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include(analytics_urls)),
+    path('', include(core_urls)),
 ]
