+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = analytics_project.settings
+DJANGO_SETTINGS_MODULE = panalytics.settings
 python_files = tests.py test_*.py *_tests.py
